Separate class definitions in SDLmain to header file
